backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0

# MongoDB
pymongo==4.6.1
dnspython==2.4.2

# Face Recognition
face-recognition==1.3.0
numpy==1.26.2
Pillow==10.1.0

# Cloudinary
cloudinary==1.37.0

# Utilities
Werkzeug==3.0.1
gunicorn==21.2.0
python-dateutil==2.8.2
requests==2.32.5
